+++ packages/dds/legacy-dds/src/test/array/arrayFuzzTests.spec.ts
@@ -9,7 +9,7 @@ import { createDDSFuzzSuite } from "@fluid-private/test-dds-utils";
 import { describe } from "mocha";
 
 import { _dirname } from "./dirname.cjs";
-import { baseSharedArrayModel } from "./fuzzUtils.js";
+import { baseSharedArrayModel, eventEmitterForFuzzHarness } from "./fuzzUtils.js";
 
 describe("SharedArray fuzz", () => {
 	createDDSFuzzSuite(baseSharedArrayModel, {
@@ -24,7 +24,10 @@ describe("SharedArray fuzz", () => {
 			numOpsBeforeAttach: 5,
 			rehydrateDisabled: true,
 		},
+		rollbackProbability: 0,
 		defaultTestCount: 50,
 		saveFailures: { directory: path.join(_dirname, "../../src/test/results") },
+		skip: [9, 15],
+		emitter: eventEmitterForFuzzHarness,
 	});
 });