+++ packages/runtime/container-runtime/src/blobManager/blobManager.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { bufferToString, createEmitter, stringToBuffer } from "@fluid-internal/client-utils";
+import { createEmitter } from "@fluid-internal/client-utils";
 import {
 	AttachState,
 	type IContainerStorageService,
@@ -41,12 +41,13 @@ import {
 	LoggingError,
 	MonitoringContext,
 	PerformanceEvent,
+	UsageError,
 	createChildMonitoringContext,
 	wrapError,
 } from "@fluidframework/telemetry-utils/internal";
 import { v4 as uuid } from "uuid";
 
-import { IBlobMetadata } from "../metadata.js";
+import { isBlobMetadata } from "../metadata.js";
 
 import {
 	getStorageIds,
@@ -148,7 +149,6 @@ interface PendingBlob {
 	attached?: boolean;
 	acked?: boolean;
 	abortSignal?: AbortSignal;
-	stashedUpload?: boolean;
 }
 
 export interface IPendingBlobs {
@@ -170,16 +170,6 @@ interface IBlobManagerInternalEvents {
 	handleAttached: (pending: PendingBlob) => void;
 	processedBlobAttach: (localId: string, storageId: string) => void;
 }
-
-const stashedPendingBlobOverrides: Pick<
-	PendingBlob,
-	"stashedUpload" | "storageId" | "minTTLInSeconds" | "uploadTime"
-> = {
-	stashedUpload: true,
-	storageId: undefined,
-	minTTLInSeconds: undefined,
-	uploadTime: undefined,
-} as const;
 
 export const blobManagerBasePath = "_blobs" as const;
 
@@ -213,8 +203,7 @@ export class BlobManager {
 	 */
 	private readonly opsInFlight: Map<string, Set<string>> = new Map();
 
-	private readonly sendBlobAttachOp: (localId: string, storageId?: string) => void;
-	private stopAttaching: boolean = false;
+	private readonly sendBlobAttachOp: (localId: string, storageId: string) => void;
 
 	private readonly routeContext: IFluidHandleContext;
 	private readonly storage: Pick<IContainerStorageService, "createBlob" | "readBlob">;
@@ -226,9 +215,6 @@ export class BlobManager {
 	private readonly isBlobDeleted: (blobPath: string) => boolean;
 	private readonly runtime: IBlobManagerRuntime;
 	private readonly localBlobIdGenerator: () => string;
-	private readonly pendingStashedBlobs: Map<string, Promise<ICreateBlobResponse | void>> =
-		new Map();
-	public readonly stashedBlobsUploadP: Promise<(void | ICreateBlobResponse)[]>;
 
 	private readonly createBlobPayloadPending: boolean;
 
@@ -247,7 +233,7 @@ export class BlobManager {
 		 * knowledge of which they cannot request the blob from storage. It's important that this op is sequenced
 		 * before any ops that reference the local ID, otherwise, an invalid handle could be added to the document.
 		 */
-		sendBlobAttachOp: (localId: string, storageId?: string) => void;
+		sendBlobAttachOp: (localId: string, storageId: string) => void;
 		// Called when a blob node is requested. blobPath is the path of the blob's node in GC's graph.
 		// blobPath's format - `/<basePath>/<blobId>`.
 		readonly blobRequested: (blobPath: string) => void;
@@ -267,7 +253,6 @@ export class BlobManager {
 			blobRequested,
 			isBlobDeleted,
 			runtime,
-			stashedBlobs,
 			localBlobIdGenerator,
 			createBlobPayloadPending,
 		} = props;
@@ -290,48 +275,7 @@ export class BlobManager {
 			this.runtime.attachState,
 		);
 
-		// Begin uploading stashed blobs from previous container instance
-		for (const [localId, entry] of Object.entries(stashedBlobs ?? {})) {
-			const { acked, storageId, minTTLInSeconds, uploadTime } = entry;
-			const blob = stringToBuffer(entry.blob, "base64");
-			const pendingEntry: PendingBlob = {
-				blob,
-				opsent: true,
-				handleP: new Deferred(),
-				storageId,
-				uploadP: undefined,
-				uploadTime,
-				minTTLInSeconds,
-				attached: true,
-				acked,
-			};
-			this.pendingBlobs.set(localId, pendingEntry);
-
-			if (storageId !== undefined && minTTLInSeconds && uploadTime) {
-				const timeLapseSinceLocalUpload = (Date.now() - uploadTime) / 1000;
-				// stashed entries with more than half-life in storage will not be reuploaded
-				if (minTTLInSeconds - timeLapseSinceLocalUpload > minTTLInSeconds / 2) {
-					continue;
-				}
-			}
-			this.pendingStashedBlobs.set(localId, this.uploadBlob(localId, blob));
-			this.pendingBlobs.set(localId, {
-				...pendingEntry,
-				...stashedPendingBlobOverrides,
-				uploadP: this.pendingStashedBlobs.get(localId),
-			});
-		}
-
-		this.stashedBlobsUploadP = PerformanceEvent.timedExecAsync(
-			this.mc.logger,
-			{ eventName: "BlobUploadProcessStashedChanges", count: this.pendingStashedBlobs.size },
-			async () => Promise.all(this.pendingStashedBlobs.values()),
-			{ start: true, end: true },
-		).finally(() => {
-			this.pendingStashedBlobs.clear();
-		});
-
-		this.sendBlobAttachOp = (localId: string, blobId?: string) => {
+		this.sendBlobAttachOp = (localId: string, blobId: string) => {
 			const pendingEntry = this.pendingBlobs.get(localId);
 			assert(
 				pendingEntry !== undefined,
@@ -385,10 +329,6 @@ export class BlobManager {
 			acked: pending?.acked,
 			uploadTime: pending?.uploadTime,
 		});
-	}
-
-	public hasPendingStashedUploads(): boolean {
-		return [...this.pendingBlobs.values()].some((e) => e.stashedUpload === true);
 	}
 
 	public hasBlob(blobId: string): boolean {
@@ -671,18 +611,9 @@ export class BlobManager {
 		response: ICreateBlobResponseWithTTL,
 	): ICreateBlobResponseWithTTL | undefined {
 		const entry = this.pendingBlobs.get(localId);
-		if (entry === undefined && this.pendingStashedBlobs.has(localId)) {
-			// The blob was already processed and deleted. This can happen if the blob was reuploaded by
-			// the stashing process and the original upload was processed before the stashed upload.
-			this.mc.logger.sendTelemetryEvent({
-				eventName: "StashedBlobAlreadyProcessed",
-				localId,
-			});
-			return;
-		}
 
 		assert(entry !== undefined, 0x6c8 /* pending blob entry not found for uploaded blob */);
-		if ((entry.abortSignal?.aborted === true && !entry.opsent) || this.stopAttaching) {
+		if (entry.abortSignal?.aborted === true && !entry.opsent) {
 			this.mc.logger.sendTelemetryEvent({
 				eventName: "BlobAborted",
 				localId,
@@ -694,7 +625,6 @@ export class BlobManager {
 			entry.storageId === undefined,
 			0x386 /* Must have pending blob entry for uploaded blob */,
 		);
-		entry.stashedUpload = undefined;
 		entry.storageId = response.id;
 		entry.uploadTime = Date.now();
 		entry.minTTLInSeconds = response.minTTLInSeconds;
@@ -738,46 +668,35 @@ export class BlobManager {
 	 * @param metadata - op metadata containing storage and/or local IDs
 	 */
 	public reSubmit(metadata: Record<string, unknown> | undefined): void {
-		assert(!!metadata, 0x38b /* Resubmitted ops must have metadata */);
-		const { localId, blobId }: { localId?: string; blobId?: string } = metadata;
-		assert(localId !== undefined, 0x50d /* local ID not available on reSubmit */);
-		const pendingEntry = this.pendingBlobs.get(localId);
-
-		if (!blobId) {
-			// We submitted this op while offline. The blob should have been uploaded by now.
-			assert(
-				pendingEntry?.opsent === true && !!pendingEntry.storageId,
-				0x38d /* blob must be uploaded before resubmitting BlobAttach op */,
-			);
-			return this.sendBlobAttachOp(localId, pendingEntry.storageId);
+		assert(isBlobMetadata(metadata), 0xc01 /* Expected blob metadata for a BlobAttach op */);
+		const { localId, blobId } = metadata;
+		// Any blob that we're actively trying to advance to attached state must have a
+		// pendingBlobs entry. Decline to resubmit for anything else.
+		// For example, we might be asked to resubmit stashed ops for blobs that never had
+		// their handle attached - these won't have a pendingBlobs entry and we shouldn't
+		// try to attach them since they won't be accessible to the customer and would just
+		// be considered garbage immediately.
+		if (this.pendingBlobs.has(localId)) {
+			this.sendBlobAttachOp(localId, blobId);
 		}
-		return this.sendBlobAttachOp(localId, blobId);
 	}
 
 	public processBlobAttachMessage(message: ISequencedMessageEnvelope, local: boolean): void {
-		const localId = (message.metadata as IBlobMetadata | undefined)?.localId;
-		const blobId = (message.metadata as IBlobMetadata | undefined)?.blobId;
-
-		if (localId) {
-			const pendingEntry = this.pendingBlobs.get(localId);
-			if (pendingEntry?.abortSignal?.aborted) {
-				this.deletePendingBlob(localId);
-				return;
-			}
+		assert(
+			isBlobMetadata(message.metadata),
+			0xc02 /* Expected blob metadata for a BlobAttach op */,
+		);
+		const { localId, blobId } = message.metadata;
+		const pendingEntry = this.pendingBlobs.get(localId);
+		if (pendingEntry?.abortSignal?.aborted) {
+			this.deletePendingBlob(localId);
+			return;
 		}
-		assert(blobId !== undefined, 0x12a /* "Missing blob id on metadata" */);
 
-		// Set up a mapping from local ID to storage ID. This is crucial since without this the blob cannot be
-		// requested from the server.
-		// Note: The check for undefined is needed for back-compat when localId was not part of the BlobAttach op that
-		// was sent when online.
-		if (localId !== undefined) {
-			this.setRedirection(localId, blobId);
-		}
+		this.setRedirection(localId, blobId);
 		// set identity (id -> id) entry
 		this.setRedirection(blobId, blobId);
 
-		assert(localId !== undefined, 0x50e /* local ID not present in blob attach message */);
 		if (local) {
 			const waitingBlobs = this.opsInFlight.get(blobId);
 			if (waitingBlobs !== undefined) {
@@ -951,31 +870,18 @@ export class BlobManager {
 	}
 
 	/**
-	 * Similar to attachAndGetPendingBlobs, but to be used in getPendingLocalState flow.
-	 * We need to store pending blobs only if we have sent a blobAttach op but not seen an ack yet,
-	 * since we may be asked to resubmit that op when restoring the container. This lets us
-	 * avoid 0x725 by having a known pending blob in that case.
+	 * To be used in getPendingLocalState flow. Get a serializable record of the blobs that are
+	 * pending upload and/or their BlobAttach op, which can be given to a new BlobManager to
+	 * resume work.
+	 *
+	 * @privateRemarks
+	 * For now, we don't track any pending blobs since the getPendingBlobs flow doesn't enable
+	 * restoring to a state where an accessible handle has been stored by the customer (and we'll
+	 * just drop any BlobAttach ops on the ground during reSubmit). However, once we add support
+	 * for payload-pending handles, this will return the blobs associated with those handles.
 	 */
 	public getPendingBlobs(): IPendingBlobs | undefined {
-		return PerformanceEvent.timedExec(this.mc.logger, { eventName: "GetPendingBlobs" }, () => {
-			if (this.pendingBlobs.size === 0) {
-				return;
-			}
-			const blobs = {};
-
-			for (const [localId, entry] of this.pendingBlobs) {
-				if (entry.opsent && !entry.acked) {
-					blobs[localId] = {
-						blob: bufferToString(entry.blob, "base64"),
-						storageId: entry.storageId,
-						acked: entry.acked,
-						minTTLInSeconds: entry.minTTLInSeconds,
-						uploadTime: entry.uploadTime,
-					};
-				}
-			}
-			return Object.keys(blobs).length > 0 ? blobs : undefined;
-		});
+		return undefined;
 	}
 
 	/**
@@ -991,83 +897,7 @@ export class BlobManager {
 	public async attachAndGetPendingBlobs(
 		stopBlobAttachingSignal?: AbortSignal,
 	): Promise<IPendingBlobs | undefined> {
-		return PerformanceEvent.timedExecAsync(
-			this.mc.logger,
-			{ eventName: "AttachAndGetPendingBlobs" },
-			async () => {
-				if (this.pendingBlobs.size === 0) {
-					return;
-				}
-				const blobs = {};
-				const localBlobs = new Set<PendingBlob>();
-				// This while is used to stash blobs created while attaching and getting blobs
-				while (localBlobs.size < this.pendingBlobs.size) {
-					const attachHandlesP: Promise<void>[] = [];
-					for (const [localId, entry] of this.pendingBlobs) {
-						if (!localBlobs.has(entry)) {
-							localBlobs.add(entry);
-							// In order to follow natural blob creation flow we need to:
-							// 1 send the blob attach op
-							// 2 resolve the blob handle
-							// 3 wait for op referencing the blob
-							if (!entry.opsent) {
-								this.sendBlobAttachOp(localId, entry.storageId);
-							}
-							// Resolving the blob handle to let hosts continue with their operations (it will resolve
-							// original createBlob call) and let them attach the blob. This is a lie we told since the upload
-							// hasn't finished yet, but it's fine since we will retry on rehydration.
-							entry.handleP.resolve(this.getBlobHandle(localId));
-							// Array of promises that will resolve when handles get attached.
-							attachHandlesP.push(
-								new Promise<void>((resolve, reject) => {
-									stopBlobAttachingSignal?.addEventListener(
-										"abort",
-										() => {
-											this.stopAttaching = true;
-											reject(new Error("Operation aborted"));
-										},
-										{ once: true },
-									);
-									const onHandleAttached = (attachedEntry: PendingBlob): void => {
-										if (attachedEntry === entry) {
-											this.internalEvents.off("handleAttached", onHandleAttached);
-											resolve();
-										}
-									};
-									if (entry.attached) {
-										resolve();
-									} else {
-										this.internalEvents.on("handleAttached", onHandleAttached);
-									}
-								}),
-							);
-						}
-					}
-					// Wait for all blobs to be attached. This is important, otherwise serialized container
-					// could send the blobAttach op without any op that references the blob, making it useless.
-					await Promise.allSettled(attachHandlesP);
-				}
-
-				for (const [localId, entry] of this.pendingBlobs) {
-					if (stopBlobAttachingSignal?.aborted && !entry.attached) {
-						this.mc.logger.sendTelemetryEvent({
-							eventName: "UnableToStashBlob",
-							id: localId,
-						});
-						continue;
-					}
-					assert(entry.attached === true, 0x790 /* stashed blob should be attached */);
-					blobs[localId] = {
-						blob: bufferToString(entry.blob, "base64"),
-						storageId: entry.storageId,
-						acked: entry.acked,
-						minTTLInSeconds: entry.minTTLInSeconds,
-						uploadTime: entry.uploadTime,
-					};
-				}
-				return Object.keys(blobs).length > 0 ? blobs : undefined;
-			},
-		);
+		throw new UsageError("attachAndGetPendingBlobs is no longer supported");
 	}
 }