+++ examples/service-clients/azure-client/external-controller/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluid-example/app-integration-external-controller
 
+## 2.52.0
+
+Dependency updates only.
+
 ## 2.51.0
 
 Dependency updates only.