+++ packages/framework/aqueduct/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluidframework/aqueduct
 
+## 2.52.0
+
+Dependency updates only.
+
 ## 2.51.0
 
 ### Minor Changes