+++ packages/service-clients/tinylicious-client/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluidframework/tinylicious-client
 
+## 2.52.0
+
+Dependency updates only.
+
 ## 2.51.0
 
 Dependency updates only.